Added SurfaceColorLayer.
Added a dedicated layer class to simulate what was done in the
DimGlobeSurface example. This class is constructed with a Color value
(defaults to black if none is specified) and is used to create a
surface-image that is stretched across the entire globe. This is
particularly useful if you want to create a Dimming effect by setting
this color to black, and playing around with the layer's opacity value
(or the image's opacity for that matter).