All for property references in XML files
When deploying Worldwind within an isolated environments with isloated
WMS servers, there is no way to centralize the definition of the server
name, instead, each of the layer XML files have to be updated.
This change allows any of the WW XML file to contain a property
reference, ${property.name}, and have these properties resolved when the
text is retrieved. Properties can be defined in three different ways;
1)worldwind.xml <Property>, 2) Java system property (-Dproperty.name),
or 3) using the legacy WW properties file.