Updated WWXML to support configuration of OutputKeys to format XML
Change WWXML to allow for configuring the OutputKeys that control the
output of the XML created from a Transformer. This is primarily to allow
for pretty printing XML docs.
Also updated RestorableSupport to allow configuring the XML produced by
this class independantly from WWXML. This allows for the XML created to
be embedded into other XML documents by disabling the xml-declaration.
Updated other classes that was directly creating Transformers and use
the WWXML.createTransformer() to allow for global configuration of XML
generation.