Small ScreenSelector bug-fix.
The ScreenSelector class should not fire a selection event when a
single-click is done on the map. It should only fire events when
clicking and dragging a rectangle. The reason for this is that the
single-click events will not contain any objects if it originates from
the ScreenSelector. This can cause odd behaviour in terms of selecting
an object and then deselecting everything. Thus, single-click select
events should only be generated by the WorldWind canvas itself.