Terrain-profile adapter.
Added a new class called TerrainProfileAdapter to allow a
TerrainProfileLayer to be updated from the MeasurementTool. When a new
control-point is added or an existing control-point is moved, the
adapter will forward the list of points to the TerrainProfileLayer so
that the new terrain-profile is displayed in the on-screen window.