Removed embedded Jackson sources.
Removed the embedded Jackson sources. Instead, we now rely on the
jackson-core-asl.jar file (version 1.9.13) as an external dependency.
Various changes were made to the ant-scripts to include the new library
on the class-path. The NetBeans and IntelliJ configurations were also
changed to reference the new dependency. The Jackson license file was
also added.