Allow for debug-builds.
Modified the build.gradle file to allow for debug-builds. This is
accomplished by specifying the property -PdebugBuild=true on the
command-line. The dependency-versions was also moved out to the ext
property clause and is now referenced in the dependencies block. A group
and description was added to the extensionsJar-task so that it correctly
falls under the build-group with the correct description when invoking
`gradle tasks`.