MacOS surface-scale fix.
According to @pcmehlitz, if we update to the latest version of JOGL and
we do not explicitly set the surface-scale factor to 1.0 in the
different constructors of WorldWindowGLCanvas, we will end up with a
surface-scale factor of 2.0 on MacOS machines with high-DPI displays. To
quote his original commit-message:

"there is one more serious issue that manifests at least on OX X machines
with high DPI (e.g. Macbook Pros with retina display). If we do not
explicitly request pixel scale factors of 1.0 in the
worldwind.awt.WorldWindowGLCanvas, we end up with factors of 2.0 after
the GLCanvas is realized, which then causes all AWTInputHandler mouse
callbacks to receive MouseEvents with wrong coordinates (0.5 of what
they should be), which in turn causes operations that compute world
coordinates from screen coordinates to fail (e.g. click-to-re-center)."

See: https://github.com/pcmehlitz/WorldWindJava/commit/3c8322d18a8db33cd219af35c2d0d98222706a8e