Default firstLevelOffset in MercatorTileUrlBuilder.
Modified the MercatorTileUrlBuilder class by adding a private constant
integer to represent the default first-level-offset value. In our case
this value is 3. A protected constructor was added that initialized the
firstLevelOffset field to this private constant integer. This allows us
to not to have to specify this value each time we instantiate a subclass
of MercatorTileUrlBuilder.