Refactored OSMMapnikLayer.
Refactored the OSMMapnikLayer class to make use of the modified
superclass constructor. Thus, we no longer have our own makeLevels()
method, but instead we call the super-class constructor, relying on the
makeLevels() method there to correctly populate the AVList parameters.
The internal URLBuilder class was also modified to extend from
MercatorTileUrlBuilder instead of TileUrlBuilder. This means that the
getMercatorURL() method had to be implemented.