Refactored OSMCycleMapLayer.
Modified the OSMCycleMapLayer class to also make use of the modified
superclass constructor. This is basically the same refactor strategy
that was applied to the OSMMapnikLayer. However, the difference here is
that we now have an addtional getter and setter method to get/set the
APIKey respectively. The reason for this is that if you want to use the
OSM-Cycle-Map from thunderforest you need to supply a valid API-key or
else you will get tiles with watermarks. By adding the getter and setter
methods we can now configure this layer in the worldwind.layers.xml file
to have the correct API-key at runtime.