Throwable getMessage() null-check in GDALUtils.
Fixed a bug in the static initialization block of GDALUtils where we
catch a Throwable and attempt to log it. The throwable's message that we
get via getMessage() might be null. So, a null-check was added that
falls back to a blank string for the message if getMessage() is null.