Property to control whether we allow 2D-pitch.
Added a property called Allow2DPitch to worldwind.xml that controls
whether pitching forward or backwards for a 2D Globe is allowed or not.
The entry for this property was also added to the AVKey class to allow
for easy reference. Modified the BasicViewPropertyLimits class by adding
the is2DGlobe() check back as well as adding a new utility method called
allow2DPitch(). This method together with the is2DGlobe() method is
called to determine whether we allow the pitch to be set in the
limitPitch() method. The default value for the new property is "false"
to ensure that default upstream behavior is kept.