Latitude normalization fix for angles > 180-deg.
Made a bugfix to the `Angle` class' normalizedDegreesLatitude() method.
The the latitude normalization didn't take the equator-crosses into
account when determining the resulting angle's sign. This fix was
spotted by @Sufaev and was taken from the Android version of WorldWind.
A couple of unit tests were also added to the `AngleTest` class to
ensure that the edge-cases are correctly accounted for.