Fixed dependencies declarations.
The JOGL, Gluegen and Jackson dependencies should be declared as
"compile" dependencies, whereas the gdal dependency should be an
implementation dependency. The reason for this is that clients that
depend on WorldWind should receive JOGL, Gluegen and Jackson as
transitive dependencies. On the other hand, they should not receive GDAL
as a transitive dependency due to the fact that GDAL can be included as
a system dependency.