Gradle task to build MilStd2525 PNGs.
Added two new Gradle tasks called processMilStd2525SVGs and
milStd2525Zip respectively. The processMilStd2525SVGs is a replacement
for the compileMilStd2525 ant-task that is present in the build.xml
file and is basically responsible for processing the MilStd2525 SVG
files to produce the corresponding PNG files. The milStd2525Zip task
depends on the output of the processMilStd2525SVGs to produce a final
zip file containing the PNGs.