Cosmetic fixes to build.gradle.
Modified some ${}-clauses by removing the curly braces as this is not
needed in this file. The doLast lambdas of jar and extensionsJar was
moved to be inside their respective tasks. This organizes the tasks a
bit better in that the tasks contain everything pertaining to their
execution inside their bodies.