Fixes to build.xml file.
To run the ant runLayerManager task we do not need to setup a GDAL
environment as the runLayerManager task does not use any GDAL
functionality. Thus, a small cleanup was done to this task by removing
the GDAL jars from the classpath as well as removing the GDAL specific
environment variables.