Allow system GDAL to be used in Gradle builds.
Modified the build.gradle file by adding an entry in the ant-block to
load either the Windows or Linux GDAL properties file based on whether
we are running in Windows or not. The dependencies clause was also
modified by adding a check to see if the "systemGDAL" property was
specified as a property on the command-line during a build. If this was
the case then the GDAL jar on the system will be used during the build,
else the default published one will be used from the maven-central or
jcenter repositories.