Decimal separator bugfix in CoordTest.
Made a small bugfix in the CoordTest class. In particular, the isClose()
method should have been called with a limit value of 0.0002 instead of a
value of 000020 (forgot to add the decimal separator). Some other
cleanups were also done by removing unnecessary zeros at the end of
some decimal numbers.